requirements.txt
streamlit>=1.20.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.2.0
tensorflow>=2.13.0
plotly>=5.15.0
scikit-learn>=1.2.0
tf2onnx>=1.15.0
onnxruntime>=1.15.0
matplotlib>=3.7.0
huggingface_hub
